elastic: catch 404 on delete and index without deleting
Catch errors on delete. It's sometimes expected to get a 404.
we may be seeing a delete message from the API for a data item we have never seen before

We don't need to delete before doing an index - index will replace too.

Also added a TODO to a code coment.